+++ NEWS.md
@@ -6,6 +6,7 @@
 ## Features
 * `tax_sort` for sorting taxa in tax_table and otu_table by several name or abundance options
 * `tax_transform` can take a rank argument, to perform aggregation (internally using tax_agg) and transformation with one function, and record the results. **This is now the recommended usage!**
+* `tax_transform` new transformation = "binary" can convert to presence/absence data (used by `tax_sort` for by = "prev")
 
 
 # microViz 0.5.0 - "hot maps"